deployment/requirements.txt
Flask
torch
torchvision
numpy
opencv-python
Pillow
joblib
clip